# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin-Panel-Api",
        "description": "Laravel Api",
        "contact": {
            "name": "Hamed",
            "email": "h.akbarzadh@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/admin/product-category/destroy/{productCategory}": {
            "delete": {
                "tags": [
                    "Product Category"
                ],
                "summary": "Delete Product Category",
                "description": "Delete Product Category",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "productCategory",
                        "in": "path",
                        "description": "Id of Product Category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/product/store": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Store Product in DB",
                "description": "Store Product in DB",
                "operationId": "Productstore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "introduction",
                                    "image",
                                    "status",
                                    "tags",
                                    "price",
                                    "brand_id",
                                    "category_id",
                                    "published_at"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Product name"
                                    },
                                    "introduction": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Product introduction"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Product image"
                                    },
                                    "tags": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Product tags"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "14,500,000"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "0 Or 1"
                                    },
                                    "brand_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "brand ID"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "Category ID"
                                    },
                                    "published_at": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "published_at"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/product/show/{product}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get Product Detail",
                "description": "Get Product Detail",
                "operationId": "Productshow",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Id of Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/product/update/{product}": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update Prorduct in DB",
                "description": "Update Prorduct in DB",
                "operationId": "productUpdate",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Id of product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "introduction",
                                    "image",
                                    "status",
                                    "tags",
                                    "price",
                                    "brand_id",
                                    "category_id",
                                    "published_at"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Product name"
                                    },
                                    "introduction": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Product introduction"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Product image"
                                    },
                                    "tags": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test Product tags"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "14,500,000"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "0 Or 1"
                                    },
                                    "brand_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "brand ID"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "Category ID"
                                    },
                                    "published_at": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "published_at"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/product/destroy/{product}": {
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete Product",
                "description": "Delete Product",
                "operationId": "productDestroy",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Id of Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/product/get-product/{number}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get Any Product Detail",
                "description": "Get Any Product Detail",
                "operationId": "get-product",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Id of Product Category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/product/status/{product}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Change Status",
                "description": "Change Status",
                "operationId": "productStatus",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Id of Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}